# backend/gwisi40server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ISI 4.0 server API",
        "title": "ISI 4.0 server API",
        "contact": {
            "name": "ICTS",
            "url": "http://www.grupoicts.com.br"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/device": {
            "post": {
                "description": "Creates a device",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devices"
                ],
                "summary": "Creates a device",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Add device",
                        "name": "type",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modelsswagger.Device"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Device"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "Missing"
                        }
                    },
                    "401": {
                        "description": "Request not authorized"
                    },
                    "404": {
                        "description": "No\t\tprotocol, type or PLC found or invalid values for name, port, memory block, class, instance or attribute"
                    }
                }
            }
        },
        "/api/v1/device/{id}": {
            "get": {
                "description": "Get device by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devices"
                ],
                "summary": "Get device by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Device ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Device"
                        }
                    },
                    "401": {
                        "description": "Request not authorized"
                    },
                    "404": {
                        "description": "No\t\t\tdevice found"
                    }
                }
            },
            "delete": {
                "description": "Deletes a device",
                "tags": [
                    "Devices"
                ],
                "summary": "Deletes a device",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Device ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Device"
                        }
                    },
                    "401": {
                        "description": "Request not authorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "Cannot"
                        }
                    }
                }
            }
        },
...
        "/api/v1/devices": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Get all devices",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devices"
                ],
                "summary": "Get all devices",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Device"
                        }
                    },
                    "401": {
                        "description": "Request not authorized"
                    },
                    "404": {
                        "description": "No\tdevices\tfound"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Device": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "integer"
                },
                "attribute": {
                    "type": "integer"
                },
                "class": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "instance": {
                    "type": "integer"
                },
                "memoryBlock": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "plc": {
                    "$ref": "#/definitions/models.PLC"
                },
                "protocol": {
                    "$ref": "#/definitions/models.Protocol"
                },
                "type": {
                    "$ref": "#/definitions/models.Type"
                }
            }
        },
...
        "models.PLC": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Protocol": {
            "type": "object",
            "properties": {
                "alias": {
                    "type": "string"
                },
                "config": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "md5": {
                    "type": "string"
                },
                "protocol": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
...
        "models.Type": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "modelsswagger.Device": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "attribute": {
                    "type": "string"
                },
                "class": {
                    "type": "string"
                },
                "instance": {
                    "type": "string"
                },
                "memoryBlock": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "plcid": {
                    "type": "string"
                },
                "protocolID": {
                    "type": "string"
                },
                "typeID": {
                    "type": "string"
                }
            }
        },
...
    }
}